requirements.txt
randaugment==1.0.2
opencv-python~=3.4.7.28
matplotlib~=3.3.2
numpy~=1.18.5
tqdm~=4.50.2
boto3~=1.16.46
Pillow~=7.2.0
sklearn~=0.0
scikit-learn~=0.23.2
torch~=1.7.1+cu101
torchvision~=0.8.2+cu101
requests~=2.24.0
tensorflow-gpu~=2.3.2
Keras~=2.4.3